Login.py
A coding project Designed for SWOSU Coding Project, allows user to login, or create a new account. All passwords are salted and hashed. It prints the salt and hash once logged in (in order to show the salt and hashes are working). 

This is one half of a code, another student is working on the tkiner GUI, and will be combined to make an application for later on.